datasets: move image/.txt pairing helper from paleography
